kitten_exhibition/requirements.txt
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
coverage==7.6.1
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
flake8==7.1.1
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
packaging==24.1
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
pytest-mock==3.14.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
setuptools==75.1.0
soupsieve==2.6
sqlparse==0.5.1
uritemplate==4.1.1
mysqlclient==2.2.4
django-mysql==4.12.0